fixed rendering bug in LabCalendars
